Functional tests now run as a standalone program
